Upgrade to ember v1.12.0 and ember-cli v0.2.7
