+++ .jshintrc
@@ -2,7 +2,8 @@
   "predef": [
     "document",
     "window",
-    "-Promise"
+    "-Promise",
+    "moment"
   ],
   "browser": true,
   "boss": true,
