+++ index.js
@@ -8,6 +8,7 @@ module.exports = {
     this._super.included(app);
     this.app.import(app.bowerDirectory + '/moment/moment.js');
     this.app.import(app.bowerDirectory + '/bootstrap-daterangepicker/daterangepicker.js');
+    this.app.import(app.bowerDirectory + '/bootstrap/dist/css/bootstrap.css');
     this.app.import(app.bowerDirectory + '/bootstrap-daterangepicker/daterangepicker-bs3.css');
   }
 };
