+++ tests/.jshintrc
@@ -6,29 +6,8 @@
     "setTimeout",
     "$",
     "-Promise",
-    "QUnit",
     "define",
     "console",
-    "equal",
-    "notEqual",
-    "notStrictEqual",
-    "test",
-    "asyncTest",
-    "testBoth",
-    "testWithDefault",
-    "raises",
-    "throws",
-    "deepEqual",
-    "start",
-    "stop",
-    "ok",
-    "strictEqual",
-    "module",
-    "moduleFor",
-    "moduleForComponent",
-    "moduleForModel",
-    "process",
-    "expect",
     "visit",
     "exists",
     "fillIn",
@@ -39,8 +18,6 @@
     "findWithAssert",
     "wait",
     "DS",
-    "isolatedContainer",
-    "startApp",
     "andThen",
     "currentURL",
     "currentPath",
