+++ tests/dummy/app/templates/application.hbs
@@ -1,3 +1,8 @@
 <h2 id="title">Welcome to Ember.js</h2>
 
+<div>
+  {{date-range-picker}}
+</div>
+
+
 {{outlet}}
