+++ tests/unit/components/date-range-picker-test.js
@@ -1,21 +1,19 @@
-import {
-  moduleForComponent,
-  test
-} from 'ember-qunit';
+import { moduleForComponent, test } from 'ember-qunit';
 
-moduleForComponent('date-range-picker', 'DateRangePickerComponent', {
-  // specify the other units that are required for this test
-  // needs: ['component:foo', 'helper:bar']
+moduleForComponent('date-range-picker', 'Unit | Component | date range picker', {
+  // Specify the other units that are required for this test
+  // needs: ['component:foo', 'helper:bar'],
+  unit: true
 });
 
-test('it renders', function() {
-  expect(2);
+test('it renders', function(assert) {
+  assert.expect(2);
 
-  // creates the component instance
+  // Creates the component instance
   var component = this.subject();
-  equal(component._state, 'preRender');
+  assert.equal(component._state, 'preRender');
 
-  // appends the component to the page
-  this.append();
-  equal(component._state, 'inDOM');
+  // Renders the component to the page
+  this.render();
+  assert.equal(component._state, 'inDOM');
 });
